Add option for custom validation error status (#178)
* Add option to use custom response status for validation errors.

Add option to specify a custom response status code for validation errors
on a global level, that'll be used for all endpoints. But, enable overriding
the status code for validation errors on the level of individual endpoints
too.

* Add validation error status code and data model to responses automatically.

Make sure that if there is a custom status code specified for validation
errors, it's used in the response spec instead of a hard-coded status code.