Check presence of media before attempting to parse it
In the falcon plugin, if we use `.data` or `.text`, then the plugin still tries to parse our response from media. Even though `.media` is not present. This PR changes the approach of parsing, we only try to parse the response model if either `.data` or `.text` is not set and `.media` is.