We have pydantic V2 installed in our environemnt, but have been using `from pydantic.v1 import ...`.  With the new changes to support pydantic V2 we are wanting to upgrade Spectree to latest and begin writing new serializers, however, we are not ready to update all of our current models to pydantic be pydantic v2 BaseModels.
Currently if a ValidationError is raised, the falcon plug in does not
catch it, since ValidationError in the plugin is coming from pydantic,
not pydantic.v1.

Since you are aliasing the pydantic.v1.ValidationError as
InternalValidationError, catching it as well allows us to continue
to support our v1 models while we migrate to v2 models.

Looking at the codebase, this looks like a similar issue in the other
plugins.  Let me know if you would like me to submit this update for
them as well.  I'm not sure where you would like a test for this in the
project.